Added use of inheritance for enemies
