+++ Assets/Scripts/EnemySystem/EnemyBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a3a5559db094f7241a7f9db9d3b5b2fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
